[IMP] hr_attendance: make use of new pivot/graph feature
Changes the previous attendance report override to use the new function
instead.

TaskId-2764844
